pkg/sandbox: take resolv.conf creation
