+++ server/utils_test.go
@@ -13,47 +13,6 @@ import (
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
-const (
-	defaultDNSPath = "/etc/resolv.conf"
-	testDNSPath    = "fixtures/resolv_test.conf"
-	dnsPath        = "fixtures/resolv.conf"
-)
-
-func TestParseDNSOptions(t *testing.T) {
-	testCases := []struct {
-		Servers, Searches, Options []string
-		Path                       string
-		Want                       string
-	}{
-		{
-			[]string{},
-			[]string{},
-			[]string{},
-			testDNSPath, defaultDNSPath,
-		},
-		{
-			[]string{"cri-o.io", "github.com"},
-			[]string{"192.30.253.113", "192.30.252.153"},
-			[]string{"timeout:5", "attempts:3"},
-			testDNSPath, dnsPath,
-		},
-	}
-
-	for _, c := range testCases {
-		if err := parseDNSOptions(c.Servers, c.Searches,
-			c.Options, c.Path); err != nil {
-			t.Error(err)
-		}
-
-		expect, _ := ioutil.ReadFile(c.Want) // nolint: errcheck
-		result, _ := ioutil.ReadFile(c.Path) // nolint: errcheck
-		if string(expect) != string(result) {
-			t.Errorf("expect %v: \n but got : %v", string(expect), string(result))
-		}
-		os.Remove(c.Path)
-	}
-}
-
 func TestMergeEnvs(t *testing.T) {
 	configImage := &v1.Image{
 		Config: v1.ImageConfig{
